Added Buttons and Static Title Card
Added in a few buttons and the title card, still need to add animations for the title card, rest of the buttons, and lore scene. Button Assets need to be resized, so don't use the ones in the Google Drive folder! Everything else should be good if you guys want to add anything yourself